+++ src/app/query-panel/query-panel.component.ts
@@ -1,4 +1,4 @@
-import {AfterViewChecked, ChangeDetectorRef, Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
+import {AfterViewChecked, Component, computed, effect, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {NgForm} from '@angular/forms';
 import {DataService} from '../singelton-services/data.service';
 import {propEach} from '@turf/meta';
@@ -7,7 +7,7 @@ import centroid from '@turf/centroid';
 import {getCoord} from '@turf/invariant';
 
 import {OhsomeApiMetadataProviderService} from '../oshdb/ohsome-api-metadata-provider.service';
-import {Feature, Polygon, GeoJsonProperties} from 'geojson';
+import {Feature, GeoJsonProperties, Polygon} from 'geojson';
 import {environment} from '../../environments/environment';
 import {BoundarySelectInputComponent} from '../shared/components/boundary-select-input/boundary-select-input.component';
 import {BoundaryInputComponent} from '../shared/components/boundary-input/boundary-input.component';
@@ -22,10 +22,10 @@ import {Subscription} from 'rxjs';
 import bboxPolygon from '@turf/bbox-polygon';
 
 @Component({
-    selector: 'app-query-panel',
-    templateUrl: './query-panel.component.html',
-    styleUrls: ['./query-panel.component.css'],
-    standalone: false
+  selector: 'app-query-panel',
+  templateUrl: './query-panel.component.html',
+  styleUrls: ['./query-panel.component.css'],
+  standalone: false
 })
 export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy {
 
@@ -34,7 +34,11 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
   @ViewChild('bsi', {static: false})
   mapInput: BoundarySelectInputComponent | BoundaryInputComponent;
 
-  public hashParams: URLSearchParams;
+  // settings from hash
+  backendParamSignal = computed(() => {
+    return this.urlHashParamsProviderService.currentHashParams().get('backend') ?? 'ohsomeApi';
+  });
+  public readonly hashParams: URLSearchParams;
 
   // default map settings
   public maskPoly;
@@ -54,7 +58,9 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
   private _selectedNames: string[] = [];
 
   public activeBackend: 'ohsomeApi' | 'oqtApi' = 'ohsomeApi';
+
   private formChangesSubscription: Subscription;
+
 
   constructor(
     private dataService: DataService,
@@ -62,8 +68,14 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
     public oqtApiMetadataProviderService: OqtApiMetadataProviderService,
     private urlHashParamsProviderService: UrlHashParamsProviderService,
     private osmBoundaryProviderService: OsmBoundaryProviderService,
-    private ref: ChangeDetectorRef
   ) {
+
+    // react on updates in the URLHashParamsProviderService
+    effect(() => {
+      const newBackend = this.backendParamSignal();
+      this.setWhichApi(newBackend);
+    });
+
 
     const spatialExtent = this.ohsomeApiMetadataProviderService
         .getOhsomeMetadataResponse()
@@ -82,12 +94,13 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
 
     //precedence: hashParams over environment over default
 
-    // settings from hash
-    this.hashParams = urlHashParamsProviderService.getHashURLSearchParams();
+    // settings from URL hashparams
+    this.hashParams = this.urlHashParamsProviderService.getHashURLSearchParams();
 
-    // settings from hash: activate the right query panel
+    // settings from URL hashparams: activate the correct query panel (ohsome or quality API)
     const backendValue = this.hashParams.get('backend');
     this.activeBackend = (backendValue === 'ohsomeApi' || backendValue === 'oqtApi') ? backendValue : 'ohsomeApi';
+
     // settings from hash: map setttings for ohsomeApi AND oqtApi
     this.bboxes = Utils.getFromParamsOrDefault(this.hashParams, 'bboxes', Utils.loadEnv('bboxes', this.bboxes));
     this.bcircles = Utils.getFromParamsOrDefault(this.hashParams, 'bcircles', Utils.loadEnv('bcircles', this.bcircles));
@@ -119,11 +132,11 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
           }
         }
       });
-  }
+
+  } // constructor end
 
   onChangeIndicatorCoverages($event: Userlayer[]) {
-    console.log("Got changes from OQT Panel")
-    console.log($event)
+
     //show additional data on the maps (e.g. coverage of comparison data in OSMAnalysis tab for specific indicators)
     /*
     1. listen to Output from oqt-panel indicator (activated indicator having GeoJSON coverage geom)
@@ -133,16 +146,14 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
 
     // Note: changing a single property of an @Input Object doesn't trigger change detection, so updating the whole
     // mapOptions Object is necessary
-    this.mapOptions = {...this.mapOptions, userDefinedPolygonLayers: $event}
+    this.mapOptions = {...this.mapOptions, userDefinedPolygonLayers: $event};
   }
 
   ngOnInit() {
+    // runs on every form change
     this.formChangesSubscription = this.form.form.valueChanges.subscribe(formValue => {
       const permalinkParams = this.getPermalinkParamsFromFormValues(formValue);
-      console.log("INIT  QueryPanel permalinkparams", permalinkParams);
-      this.urlHashParamsProviderService.updateHashParams(permalinkParams);
-      //needed to update selected name labels when areas are unselected in the map
-      this.ref.markForCheck();
+      this.urlHashParamsProviderService.setHashParams(permalinkParams);
     })
   }
 
@@ -254,11 +265,10 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
 
 
       // transform attribute-completeness--attributes
-      if (permalinkParams["attribute-completeness--attributes"]){
+      if (permalinkParams["attribute-completeness--attributes"]) {
         permalinkParams["attribute-completeness--attributes"] = permalinkParams["attribute-completeness--attributes"].join(',');
       }
     }
-
 
 
     return permalinkParams;
@@ -269,12 +279,11 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
     console.log('Form Value', this.form.value);
     const permalinkParams = this.getPermalinkParamsFromFormValues(this.form.value);
     console.log("ONSUBMIT QueryPanel permalinkparams", permalinkParams);
-    this.urlHashParamsProviderService.updateHashParams(permalinkParams);
     this.dataService.pushFormValues(this.form.value, this._boundaryType);
   }
 
-  setWhichApi(activeApi: 'ohsomeApi' | 'oqtApi') {
-    this.activeBackend = activeApi;
+  setWhichApi(activeApi: string | null): void {
+    this.activeBackend = (activeApi === 'ohsomeApi' || activeApi === 'oqtApi')? activeApi : 'ohsomeApi';
     if (activeApi === 'oqtApi' && this.boundaryType === 'bcircle') {
       this.boundaryType = 'admin';
     }
@@ -309,4 +318,5 @@ export class QueryPanelComponent implements OnInit, AfterViewChecked, OnDestroy 
   protected readonly window = window;
   protected readonly Object = Object;
   protected readonly JSON = JSON;
+  // protected readonly $localize = $localize;
 }