+++ src/app/singelton-services/url-hash-params-provider.service.mock.ts
@@ -1,0 +1,16 @@
+import createSpyObj = jasmine.createSpyObj;
+
+const UrlHashParamsProviderServiceMock = createSpyObj(
+  "UrlHashParamsProviderService",
+  [
+    "updateHashParamsStoreFromUrl",
+    "getHashURLSearchParams",
+    "setHashParams",
+    "updateHashParams",
+    "updateHashParam"
+  ]
+);
+
+UrlHashParamsProviderServiceMock.getHashURLSearchParams.and.returnValue(new URLSearchParams({backend: "ohsomeApi"}));
+
+export default UrlHashParamsProviderServiceMock;