+++ src/environments/environment.ts
@@ -12,13 +12,13 @@ export const environment = {
   //oqtApi (ohsome quality analyst)
   // oqtApiRootUrl: 'https://api.quality.ohsome.org/v1',
   oqtApiRootUrl: 'https://api.quality.ohsome.org/v1-test',
-  // oqtApiRootUrl: 'http://127.0.0.1:8080',
+  // oqtApiRootUrl: 'http://127.0.0.1:8000',
   // core, experimental, all, ...
-  oqtApiProject: 'bkg',
+  oqtApiProject: 'all',
 
   //the WFS endpoint which will get appended a CQL_FILTER="id" IN (number, number, ...) param and should return a GeoJSON<FeatureCollection>
-  ohsomeBoundaryWFSUrl: 'https://maps.heigit.org/ohsome/wfs?service=wfs&request=GetFeature&typeNames=ohsome:admin_world_water&outputFormat=application/json&version=2.0.0&srsName=EPSG:4326',
-  ohsomeBoundaryWMSUrl: 'https://maps.heigit.org/ohsome/service/wms',
+  ohsomeBoundaryWFSUrl: 'https://maps.heigit.org/vector/service/ohsome/wfs?service=wfs&request=GetFeature&typeNames=ohsome:admin_world_water&outputFormat=application/json&version=2.0.0&srsName=EPSG:4326',
+  ohsomeBoundaryWMSUrl: 'https://maps.heigit.org/raster/ohsome/wms',
   ohsomeBoundaryWMSLayer: 'ohsome:admin_world_water',
 
   // url to fetch an announcement text