+++ src/main.ts
@@ -1,7 +1,8 @@
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+/// <reference types="@angular/localize" />
 
-import { AppModule } from './app/app.module';
+import {AppModule} from './app/app.module';
+import {platformBrowser} from '@angular/platform-browser';
+import { $localize } from '@angular/localize/init';
 
-
-platformBrowserDynamic().bootstrapModule(AppModule)
+platformBrowser().bootstrapModule(AppModule)
   .catch(err => console.error(err));